main/postmarketos-ui-{gnome-mobile,phosh,plasma-mobile}: depend on pmos-usb-moded
Let usb-moded handle the USB gadget in GNOME Mobile, Phosh and Plasma Mobile UIs.

OpenRC and systemd are both using the patches of usb-moded to allow
switching modes through a set of scripts. usb-moded supports also
activation of systemd services depending on the mode which is the
'normal' way. If this is used, the patches are not all needed for
usb-moded, but would create different code paths for systemd and OpenRC,
which is not good for maintenance.

Co-authored-by: Clayton Craft <craftyguy@postmarketos.org>
Part-of: <https://gitlab.postmarketos.org/postmarketOS/pmaports/-/merge_requests/6531>