systemd/cups: disable examples / ipptool on all archs
Apparently this fails on other archs when cross compiling, not just
riscv64. I hit it locally when building x86_64 on aarch64. bpo is
hitting it when trying to build aarch64 on x86_64. This is kinda of a
big hammer, but I think we'll be able to drop this cups fork when abuild
systemd split func support is upstreamed, and bpo is currently blocked
trying to build changes to support our effort to upstream that abuild
patch.

Part-of: <https://gitlab.postmarketos.org/postmarketOS/pmaports/-/merge_requests/7023>