{temp,systemd}/gnome-shell-mobile: backport GIRepository 2.0 patches for GJS >= 1.85.90 support
GJS 1.86 was released with a port from
libgirepository-1.0 to libgirepository-2.0. That port
requires changes in gnome-shell. However, for now, there
is no gnome-shell-mobile based on 49, so as a hot fix,
this is a backport of those patches only.

Without these, the GNOME Shell will not even get beyond
main.js and as such is unusable even for gdm.

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>
Part-of: <https://gitlab.postmarketos.org/postmarketOS/pmaports/-/merge_requests/7061>