device-nothing-tetris: adapt to U-Boot
As a first-stage U-Boot port is now available for this device, adapt the
device package to use it.

NOTE: Currently, EFI boot is unsupported, because of this device having
to send a specific SMC command on kernel launch, if this is not done,
then the device will crash after ~30s with weird errors, and access to
peripherals shared with the Secure World (for example UFS) will be
unavailable. This is only a temporary workaround for the time being.

Signed-off-by: Igor Belwon <igor.belwon@mentallysanemainliners.org>
Part-of: <https://gitlab.postmarketos.org/postmarketOS/pmaports/-/merge_requests/7097>