main/postmarketos-ui-plasma-mobile: drop systemd unit file
This should come from the package providing the daemon run by the unit
file (and in our case, at least temporarily, from the vendored unit file
in systemd-services)

This also adds an install script for post-upgrade that makes sure that
the alias for display-manager.service is created for plasma-mobile on
existing installs where we were missing the alias definition in the
tinydm unit file.
Without this workaround, existing installs will miss the alias and we
won't be able to run things that are associated with this alias (like
stopping pbsplash.)
Note: this is a post-install script since this subpkg is new, and on
existing systems affected by the issue it resolves the subpkg would be
'installed' not 'upgraded'.

Part-of: <https://gitlab.postmarketos.org/postmarketOS/pmaports/-/merge_requests/7202>