device-samsung-gtaxlwifi: remove obsolete and invalid properties
While I haven't verified it personally, the wiki entry for this device
states that neither Weston nor Plasma Mobile work on the device,
indicating that DRM probably doesn't actually work. Furthermore, it is a
downstream port, so it working would be a miracle regardless. As for the
keyboard property, it is obsolete.

Part-of: <https://gitlab.postmarketos.org/postmarketOS/pmaports/-/merge_requests/7243>