soc-qcom-sc7280: replace with soc-qcom
With the addition of 81voltd the difference between
soc-qcom-sc7280 and soc-qcom shrank. With the unreliable
hardware en- and decoder and a workaround in soc-qcom,
it would shrink even further.

As such, it makes sense to replace the former with the
latter.

With q6voiced and a call audio wireplumber config,
soc-qcom adds two more things untestable on the two
devices currently using soc-qcom-sc7280 (Fairphone 5 and
Shiftphone 8).

Besides that, it is only a UPower.conf replacing suspend
with shutdown in case of critically low battery levels
and setting the low level from 20 % to 10 %, which makes
sense on these two devices as well.

Link: https://gitlab.postmarketos.org/postmarketOS/pmaports/-/merge_requests/7044#note_501998
Link: https://gitlab.postmarketos.org/postmarketOS/pmaports/-/merge_requests/7107#note_508712

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>
Part-of: <https://gitlab.postmarketos.org/postmarketOS/pmaports/-/merge_requests/7254>