Remove apq8064 kernel
This kernel shouldn't exist. msm8960 devices, like the expressatt,
already use the linux-* mainline kernels and the apq8064 is very related
to the msm8960. Let's deprecate this kernel and shift the devices that
depend on the apq8064 kernel onto the mainline kernels.

Part-of: <https://gitlab.postmarketos.org/postmarketOS/pmaports/-/merge_requests/7334>