device-purism-librem5: depend on bash
We package the lockdown-support.sh script from Purism's librem5-base
package. This script has a bash shebang[1] and so won't work with any
other shell. It also uses some bashisms that Ash doesn't support, so we
can't just change the shebang to be /bin/sh or such as that will break
the script.

As such, just solve this in the simplest way for now and depend on bash.
If someone feels motivated to replace the bashisms in the upstream
script they're very welcome to do so.

Closes https://gitlab.postmarketos.org/postmarketOS/pmaports/-/issues/4089

 [1]: https://source.puri.sm/Librem5/librem5-base/-/blob/07a4105dd4d96158e11d6796a25768926c7a3f30/default/lockdown-support/lockdown-support.sh#L1

Signed-off-by: Stefan Hansson <newbyte@postmarketos.org>
Part-of: <https://gitlab.postmarketos.org/postmarketOS/pmaports/-/merge_requests/7376>