main/unl0kr: do not copy initramfs files during tests a second time
Since this was introduced, we no longer create "initramfs-extra"
unconditionally. So the "extra" files will be included in the
initramfs anyway. In addition, to be able to test the initramfs-extra
split for devices that need it, we require:

* To have one of those devices available for test.
* To modify our test setup to pass a rootfs that can be mounted.

Both are long into the future, if they will ever materialize. So let's
remove this, reducing the complexity.

Signed-off-by: Pablo Correa Gómez <pabloyoyoista@postmarketos.org>
Part-of: <https://gitlab.postmarketos.org/postmarketOS/pmaports/-/merge_requests/7553>