device-qcom-sm7150: Disable splash and serial console for now
On systemd installs, pbsplash causes the device to hang during init and
reset. OpenRC does not exhibit this behavior. Let's disable the splash
screen for now to get the devices working out of the box again and look
into this again after 25.12 when we have plymouth.

The serial getty setup is broken on some devices. Not all of them have
the UART TX/RX connected in hardware, so the ttyMSM0 does not exist on
all devices. On those that have it, it works, on those that don't, it
causes logspam.

Signed-off-by: Jens Reidel <adrian@postmarketos.org>
Part-of: <https://gitlab.postmarketos.org/postmarketOS/pmaports/-/merge_requests/7509>