systemd/dbus: rebuild with proper deps
The dbus-dev package in ppc64le does not depend on systemd-dev yet, as
it was not rebuilt after the initial bootstrap of the repository.

See depends here:
* https://pkgs.postmarketos.org/package/master/postmarketos:systemd/ppc64le/dbus-dev
* https://pkgs.postmarketos.org/package/master/postmarketos:systemd/riscv64/dbus-dev

Rebuild it to fix that packages building against dbus-dev fail with:
  Package 'libsystemd', required by 'dbus-1', not found

All arches in v25.12 have the same problem, so this patch will be
backported there. Long-term we will want to fix this properly, but given
that the release is close, this is a good enough workaround with low
risk of breaking anything.

Related: pmaports issue 4156
Part-of: <https://gitlab.postmarketos.org/postmarketOS/pmaports/-/merge_requests/7543>