main/postmarketos-ui-lomiri: fix wallpaper - yet again
Just pull in postmarketos-artwork-wallpapers-gnome which provides a
gschema override that is updated when the artwork package gets updated -
and won't break every so often.

Usually that package is pulled in via an "install_if" with
postmarketos-base-ui-gnome but since we just want the artwork
subpackage, depending on base-ui-gnome would be quite much and probably
not what we want for Lomiri.

Part-of: <https://gitlab.postmarketos.org/postmarketOS/pmaports/-/merge_requests/7593>